break responsibilities into separate classes
