+++ message.txt
@@ -1,0 +1,1 @@
+some text
